feat: redis cache for catalog

+++ src/main/java/com/dtp/cosmemgt/catalog/service/AdminUomService.java
@@ -5,6 +5,7 @@ import com.dtp.cosmemgt.catalog.entity.UnitOfMeasure;
 import com.dtp.cosmemgt.catalog.dto.request.UomCreationRequest;
 import com.dtp.cosmemgt.catalog.dto.response.UomResponse;
 import com.dtp.cosmemgt.catalog.mapper.UomMapper;
+import com.dtp.cosmemgt.catalog.repository.ProductVariantRepository;
 import com.dtp.cosmemgt.catalog.repository.UomRepository;
 import com.dtp.cosmemgt.core.exception.AppException;
 import com.dtp.cosmemgt.core.exception.ErrorCode;
@@ -12,6 +13,8 @@ import lombok.AccessLevel;
 import lombok.RequiredArgsConstructor;
 import lombok.experimental.FieldDefaults;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.cache.annotation.CacheEvict;
+import org.springframework.cache.annotation.Cacheable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -23,9 +26,11 @@ import java.util.List;
 @Transactional
 @Slf4j
 public class AdminUomService {
+    private final ProductVariantRepository productVariantRepository;
     UomRepository uomRepository;
     UomMapper uomMapper;
 
+    @CacheEvict(value = "Uoms", allEntries = true)
     public UomResponse create(UomCreationRequest request) {
         UnitOfMeasure u = uomMapper.toUom(request);
 
@@ -49,20 +54,26 @@ public class AdminUomService {
                 .toList();
     }
 
+    @CacheEvict(value = "Uoms", allEntries = true)
     public UomResponse update(int uomId, UomUpdateRequest request){
         UnitOfMeasure u = uomRepository.findById(uomId)
-                .orElseThrow(() -> new AppException(ErrorCode.CATEGORY_NOT_EXISTED));
+                .orElseThrow(() -> new AppException(ErrorCode.UOM_NOT_EXISTED));
         u.setName(request.getName());
 
         return uomMapper.toUomResponse(uomRepository.save(u));
     }
 
+    @CacheEvict(value = "Uoms", allEntries = true)
     public void sftDelUom(int uomId) {
         UnitOfMeasure u = uomRepository.findById(uomId)
                 .orElseThrow(() -> new AppException(ErrorCode.UOM_NOT_EXISTED));
+        if (productVariantRepository.existsByUnitOfMeasureId(uomId)) {
+            throw new AppException(ErrorCode.UOM_IN_USE);
+        }
         uomRepository.delete(u);
     }
 
+    @CacheEvict(value = "Uoms", allEntries = true)
     public void restoreUom(int uomId) {
         int rowsAffected = uomRepository.restoreById(uomId);
         if (rowsAffected == 0) {
@@ -73,6 +84,9 @@ public class AdminUomService {
     public void hardDelUom(int uomId) {
         UnitOfMeasure u = uomRepository.getById(uomId)
                 .orElseThrow(() -> new AppException(ErrorCode.UOM_NOT_EXISTED));
+        if (productVariantRepository.existsByUnitOfMeasureId(uomId)) {
+            throw new AppException(ErrorCode.UOM_IN_USE);
+        }
         uomRepository.hardDelById(u.getId());
     }
 }
